Case 27689; updated checkout command to support branch and tag options
